requirements.txt
Flask
SQLAlchemy
PyYAML
requests
psycopg2-binary
python-dotenv

